@@ -1,6 +1,6 @@
 import { errorMessages } from './errors'
-import { IgdbItem } from '@/types'
-import { encodeQuery } from '@/helpers/search'
+import type { IgdbItem } from '../types'
+import { encodeQuery } from '../helpers/search'
 
 const queryIGDB = async (query: string): Promise<IgdbItem[]> => {
   if (query === '') {
@@ -9,7 +9,7 @@ const queryIGDB = async (query: string): Promise<IgdbItem[]> => {
 
   const encodedQuery = encodeQuery(query)
 
-  const res = await fetch(`${process.env.VUE_APP_BACKEND_URL}/api/igdb/search/${encodedQuery}`)
+  const res = await fetch(`${import.meta.env.VITE_BACKEND_URL}/api/igdb/search/${encodedQuery}`)
 
   if (!res) {
     throw new Error(errorMessages.NoConnection)
