@@ -1,20 +1,13 @@
+<script setup lang="ts">
+import Options from './Sidebar/Options.vue'
+</script>
+
 <template>
   <div class="container">
     <h1>Options</h1>
     <Options />
   </div>
 </template>
-
-<script lang="ts">
-import { defineComponent } from '@vue/runtime-core'
-import Options from './Sidebar/Options.vue'
-
-export default defineComponent({
-  components: {
-    Options
-  }
-})
-</script>
 
 <style scoped>
 .container {
