@@ -1,3 +1,135 @@
+<script setup lang="ts">
+import queryOpenLibrary from '../../../api/openlibrary'
+import TypeDropdown from './TypeDropdown.vue'
+import { BIconArrowRight, BIconArrowRepeat } from 'bootstrap-icons-vue'
+import queryLastFM from '../../../api/lastfm'
+import queryIGDB from '../../../api/igdb'
+import { tmdbMovieSearch, tmdbTVSearch } from '../../../api/tmdb'
+import type { CustomResult, SearchTypes } from '../../../types'
+import { ref } from 'vue'
+
+interface Props {
+  searchType: SearchTypes
+}
+
+const props = defineProps<Props>()
+
+const results = ref([])
+const loading = ref(false)
+const customImageData = ref({
+  title: '',
+  creator: ''
+})
+
+const emit = defineEmits([
+  'updateResults',
+  'setSearchType'
+])
+
+const searchbox = ref(null)
+
+const changeSearchType = (newType: string) => {
+  emit('setSearchType', newType)
+}
+
+const focusSearchBox = () => {
+  searchbox.value.focus()
+  searchbox.value.select()
+}
+
+const changeCustomImageData = (attr: 'title' | 'creator', value: string): void => {
+  customImageData.value[attr] = value
+  const item = results.value[0] as CustomResult
+  item[attr] = value
+}
+
+const handleSearch = async (): Promise<unknown[] | null> => {
+  if (loading.value) {
+    return null
+  }
+
+  const query = searchbox.value.value
+
+  switch (props.searchType) {
+    case 'books':
+    {
+      loading.value = true
+      const response = await queryOpenLibrary(query)
+      // TODO: needs error handling
+      if (response) {
+        results.value = response
+        emit('updateResults', response)
+      }
+      loading.value = false
+      break
+    }
+    case 'music':
+    {
+      loading.value = true
+      const response = await queryLastFM(query)
+      if (response) {
+        results.value = response
+        emit('updateResults', response)
+      }
+      loading.value = false
+      break
+    }
+    case 'games':
+    {
+      loading.value = true
+      const response = await queryIGDB(query)
+      if (response) {
+        results.value = response
+        emit('updateResults', response)
+      }
+      loading.value = false
+      break
+    }
+    case 'movies':
+    {
+      loading.value = true
+      const response = await tmdbMovieSearch(query)
+      if (response) {
+        results.value = response
+        emit('updateResults', response)
+      }
+      loading.value = false
+      break
+    }
+    case 'tv':
+    {
+      loading.value = true
+      const response = await tmdbTVSearch(query)
+      if (response) {
+        results.value = response
+        emit('updateResults', response)
+      }
+      loading.value = false
+      break
+    }
+    case 'custom':
+    {
+      console.log(query)
+      loading.value = true
+      results.value = [
+        {
+          title: customImageData.value.title,
+          creator: customImageData.value.creator,
+          imageURL: query,
+          type: 'custom'
+        }
+      ]
+      emit('updateResults', results.value)
+      loading.value = false
+      break
+    }
+    default:
+      return []
+  }
+  return []
+}
+</script>
+
 <template>
   <form
     class="add-form"
@@ -12,7 +144,7 @@
         <input
           id="searchbox"
           ref="searchbox"
-          @click="searchType === 'custom' ? this.focusSearchBox() : null"
+          @click="searchType === 'custom' ? focusSearchBox() : null"
           :placeholder="searchType === 'custom' ? 'Enter URL' : null"
         />
         <button type="submit" class="submit-button">
@@ -30,7 +162,7 @@
         <input
           type="text"
           id="custom-creator"
-          @change="changeCustomImageData('creator', $event.target.value)"
+          @change="changeCustomImageData('creator', ($event.target as HTMLInputElement).value)"
         />
       </div>
       <div class="custom-form-item">
@@ -38,156 +170,12 @@
         <input
           type="text"
           id="custom-title"
-          @change="changeCustomImageData('title', $event.target.value)"
+          @change="changeCustomImageData('title', ($event.target as HTMLInputElement).value)"
         />
       </div>
     </div>
   </form>
 </template>
-
-<script lang="ts">
-import queryOpenLibrary from '../../../api/openlibrary'
-import TypeDropdown from './TypeDropdown.vue'
-import { defineComponent } from '@vue/runtime-core'
-import { BIconArrowRight, BIconArrowRepeat } from 'bootstrap-icons-vue'
-import { SearchTypes } from './index.vue'
-import queryLastFM from '../../../api/lastfm'
-import queryIGDB from '@/api/igdb'
-import { tmdbMovieSearch, tmdbTVSearch } from '../../../api/tmdb'
-import { CustomResult } from '@/types'
-
-interface FormData {
-  results: unknown[],
-  loading: boolean,
-  customImageData: {
-    title: string,
-    creator: string
-  }
-}
-
-export default defineComponent({
-  components: {
-    TypeDropdown,
-    BIconArrowRight,
-    BIconArrowRepeat
-  },
-  props: [
-    'searchType'
-  ],
-  emits: [
-    'updateResults',
-    'setSearchType'
-  ],
-  data (): FormData {
-    return {
-      results: [] as FormData[],
-      loading: false,
-      customImageData: {
-        title: '',
-        creator: ''
-      }
-    }
-  },
-  methods: {
-    changeSearchType (newType: string) {
-      this.$emit('setSearchType', newType)
-    },
-    focusSearchBox () {
-      (this.$refs.searchbox as HTMLInputElement).focus();
-      (this.$refs.searchbox as HTMLInputElement).select()
-    },
-    changeCustomImageData (attr: 'title' | 'creator', value: string): void {
-      this.customImageData[attr] = value
-      const item = this.results[0] as CustomResult
-      item[attr] = value
-    },
-    async handleSearch (): Promise<unknown[] | null> {
-      if (this.loading) {
-        return null
-      }
-
-      const query = (this.$refs.searchbox as HTMLInputElement).value
-
-      switch (this.searchType) {
-        case SearchTypes.Books:
-        {
-          this.loading = true
-          const response = await queryOpenLibrary(query)
-          // TODO: needs error handling
-          if (response) {
-            this.results = response
-            this.$emit('updateResults', response)
-          }
-          this.loading = false
-          break
-        }
-        case SearchTypes.Music:
-        {
-          this.loading = true
-          const response = await queryLastFM(query)
-          if (response) {
-            this.results = response
-            this.$emit('updateResults', response)
-          }
-          this.loading = false
-          break
-        }
-        case SearchTypes.Games:
-        {
-          this.loading = true
-          const response = await queryIGDB(query)
-          if (response) {
-            this.results = response
-            this.$emit('updateResults', response)
-          }
-          this.loading = false
-          break
-        }
-        case SearchTypes.Movies:
-        {
-          this.loading = true
-          const response = await tmdbMovieSearch(query)
-          if (response) {
-            this.results = response
-            this.$emit('updateResults', response)
-          }
-          this.loading = false
-          break
-        }
-        case SearchTypes.TV:
-        {
-          this.loading = true
-          const response = await tmdbTVSearch(query)
-          if (response) {
-            this.results = response
-            this.$emit('updateResults', response)
-          }
-          this.loading = false
-          break
-        }
-        case SearchTypes.Custom:
-        {
-          this.loading = true
-          this.results = [
-            {
-              title: this.customImageData.title,
-              creator: this.customImageData.creator,
-              imageURL: query,
-              type: 'custom'
-            }
-          ]
-          this.$emit('updateResults', this.results)
-          this.loading = false
-          break
-        }
-        default:
-          return []
-      }
-      return []
-    }
-  }
-})
-</script>
 
 <style scoped>
 .add-form {
