@@ -2,6 +2,6 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import { store, key } from './store'
 
-const app = createApp(App)
-app.use(store, key)
-app.mount('#app')
+createApp(App)
+  .use(store, key)
+  .mount('#app')
